Allow the service to stop itself when interactive
If the service is running from the command line and it calls
```
ServiceController.Stop()
```
the program will log that the service is stopping, but does not end,
instead requiring a ctrl+c from the user. This a problem for scripts
that are intended to be run without anyone watching the command line.